+++ src/Feature/BasicContent/items/templates/BasicContent/ContentList/Content List/ContentListLimit.yml
@@ -1,8 +1,8 @@
 ﻿---
-ID: "e5b87c88-51ef-4395-bab4-d8e5eb2f2ee5"
-Parent: "c4b9313a-3670-4896-ad8c-de611938a194"
+ID: "5ad6837f-54d4-47d9-9e7a-5ec30f6b0f13"
+Parent: "826f5cb6-2237-43ae-b2af-6df0576a4d9d"
 Template: "455a3e98-a627-4b40-8035-e683a0331ac7"
-Path: /sitecore/templates/Feature/BasicContent/Rendering Parameters/Content List Rendering Parameters/Content List Properties/ContentListLimit
+Path: /sitecore/templates/Feature/BasicContent/ContentList/Content List/ContentListLimit
 SharedFields:
 - ID: "ab162cc0-dc80-4abf-8871-998ee5d7ba32"
   Hint: Type
@@ -20,11 +20,11 @@ Languages:
       Value: 20200630T200213Z
     - ID: "8cdc337e-a112-42fb-bbb4-4143751e123f"
       Hint: __Revision
-      Value: "11e7e8ab-3324-4299-9ba8-c42663b82240"
+      Value: "b7827895-4da2-406f-8556-32d800db607d"
     - ID: "badd9cf9-53e0-4d0c-bcc0-2d784c282f6a"
       Hint: __Updated by
       Value: |
-        sitecore\Admin
+        sitecore\xmc-e2e-admin@sitecore.com
     - ID: "d9cf14b1-fa16-4ba6-9288-e8a174d4d522"
       Hint: __Updated
-      Value: 20200630T200237Z
+      Value: 20220524T010213Z
