+++ src/Feature/BasicContent/items/templates/BasicContent/ContentList/Content List/ContentListTitle.yml
@@ -9,7 +9,7 @@ SharedFields:
   Value: "Single-Line Text"
 - ID: "ba3f86a2-4a1c-4d78-b63d-91c2779c1b5e"
   Hint: __Sortorder
-  Value: 100
+  Value: 50
 Languages:
 - Language: en
   Versions:
@@ -20,11 +20,11 @@ Languages:
       Value: 20200629T110351Z
     - ID: "8cdc337e-a112-42fb-bbb4-4143751e123f"
       Hint: __Revision
-      Value: "4c39ff27-3369-464a-b671-fb20dd14498d"
+      Value: "59cc2311-1faf-4f38-8326-31a1bb475949"
     - ID: "badd9cf9-53e0-4d0c-bcc0-2d784c282f6a"
       Hint: __Updated by
       Value: |
-        sitecore\Admin
+        sitecore\xmc-e2e-admin@sitecore.com
     - ID: "d9cf14b1-fa16-4ba6-9288-e8a174d4d522"
       Hint: __Updated
-      Value: 20200629T111514Z
+      Value: 20220524T010213Z
