--- src/Feature/BasicContent/items/templates/BasicContent/Rendering Parameters.yml
@@ -1,27 +1,0 @@
-﻿---
-ID: "495c6945-e34c-4d36-84ac-0c3398fee3d7"
-Parent: "2425c9e1-3b22-44b5-9da8-704ff22d388e"
-Template: "0437fee2-44c9-46a6-abe9-28858d9fee8c"
-Path: /sitecore/templates/Feature/BasicContent/Rendering Parameters
-SharedFields:
-- ID: "ba3f86a2-4a1c-4d78-b63d-91c2779c1b5e"
-  Hint: __Sortorder
-  Value: 50
-Languages:
-- Language: en
-  Versions:
-  - Version: 1
-    Fields:
-    - ID: "25bed78c-4957-4165-998a-ca1b52f67497"
-      Hint: __Created
-      Value: 20200629T082856Z
-    - ID: "8cdc337e-a112-42fb-bbb4-4143751e123f"
-      Hint: __Revision
-      Value: "7dd81126-10bf-4936-854f-8e0297b47857"
-    - ID: "badd9cf9-53e0-4d0c-bcc0-2d784c282f6a"
-      Hint: __Updated by
-      Value: |
-        sitecore\Admin
-    - ID: "d9cf14b1-fa16-4ba6-9288-e8a174d4d522"
-      Hint: __Updated
-      Value: 20211206T141356Z
