--- src/Feature/BasicContent/items/templates/BasicContent/Rendering Parameters/Content List Rendering Parameters.yml
@@ -1,30 +1,0 @@
-﻿---
-ID: "f0105115-d6c7-42b5-8793-d3f6dd215c32"
-Parent: "495c6945-e34c-4d36-84ac-0c3398fee3d7"
-Template: "ab86861a-6030-46c5-b394-e8f99e8b87db"
-Path: /sitecore/templates/Feature/BasicContent/Rendering Parameters/Content List Rendering Parameters
-SharedFields:
-- ID: "12c33f3f-86c5-43a5-aeb4-5598cec45116"
-  Hint: __Base template
-  Value: "{8CA06D6A-B353-44E8-BC31-B528C7306971}"
-- ID: "f7d48a55-2158-4f02-9356-756654404f73"
-  Hint: __Standard values
-  Value: "{2077A3DF-63C0-48EC-A050-46CC0298BCAF}"
-Languages:
-- Language: en
-  Versions:
-  - Version: 1
-    Fields:
-    - ID: "25bed78c-4957-4165-998a-ca1b52f67497"
-      Hint: __Created
-      Value: 20200630T193044Z
-    - ID: "8cdc337e-a112-42fb-bbb4-4143751e123f"
-      Hint: __Revision
-      Value: "ba751901-c40d-4740-8eab-1a53cf8591f9"
-    - ID: "badd9cf9-53e0-4d0c-bcc0-2d784c282f6a"
-      Hint: __Updated by
-      Value: |
-        sitecore\Admin
-    - ID: "d9cf14b1-fa16-4ba6-9288-e8a174d4d522"
-      Hint: __Updated
-      Value: 20200630T200254Z
