--- src/Feature/BasicContent/items/templates/BasicContent/Rendering Parameters/Content List Rendering Parameters/Content List Properties.yml
@@ -1,23 +1,0 @@
-﻿---
-ID: "c4b9313a-3670-4896-ad8c-de611938a194"
-Parent: "f0105115-d6c7-42b5-8793-d3f6dd215c32"
-Template: "e269fbb5-3750-427a-9149-7aa950b49301"
-Path: /sitecore/templates/Feature/BasicContent/Rendering Parameters/Content List Rendering Parameters/Content List Properties
-Languages:
-- Language: en
-  Versions:
-  - Version: 1
-    Fields:
-    - ID: "25bed78c-4957-4165-998a-ca1b52f67497"
-      Hint: __Created
-      Value: 20200630T193103Z
-    - ID: "8cdc337e-a112-42fb-bbb4-4143751e123f"
-      Hint: __Revision
-      Value: "c41ff62e-ef09-4de8-bcff-28a6e3081387"
-    - ID: "badd9cf9-53e0-4d0c-bcc0-2d784c282f6a"
-      Hint: __Updated by
-      Value: |
-        sitecore\Admin
-    - ID: "d9cf14b1-fa16-4ba6-9288-e8a174d4d522"
-      Hint: __Updated
-      Value: 20200630T200237Z
