--- src/Feature/BasicContent/items/templates/BasicContent/Rendering Parameters/Content List Rendering Parameters/__Standard Values.yml
@@ -1,23 +1,0 @@
-﻿---
-ID: "2077a3df-63c0-48ec-a050-46cc0298bcaf"
-Parent: "f0105115-d6c7-42b5-8793-d3f6dd215c32"
-Template: "f0105115-d6c7-42b5-8793-d3f6dd215c32"
-Path: /sitecore/templates/Feature/BasicContent/Rendering Parameters/Content List Rendering Parameters/__Standard Values
-Languages:
-- Language: en
-  Versions:
-  - Version: 1
-    Fields:
-    - ID: "25bed78c-4957-4165-998a-ca1b52f67497"
-      Hint: __Created
-      Value: 20200630T200254Z
-    - ID: "8cdc337e-a112-42fb-bbb4-4143751e123f"
-      Hint: __Revision
-      Value: "49850dc0-07b4-47ea-872e-9d19c89ca877"
-    - ID: "badd9cf9-53e0-4d0c-bcc0-2d784c282f6a"
-      Hint: __Updated by
-      Value: |
-        sitecore\Admin
-    - ID: "d9cf14b1-fa16-4ba6-9288-e8a174d4d522"
-      Hint: __Updated
-      Value: 20200630T200254Z
