+++ src/Feature/BasicContent/rendering/Views/Shared/Components/SitecoreComponent/ContentList.cshtml
@@ -1,25 +1,25 @@
 ﻿@model Mvp.Feature.BasicContent.Models.ContentList
 
 
-@if (!string.IsNullOrEmpty(Model.Title))
+@if (!string.IsNullOrEmpty(Model.Title.Value))
 {
     <div class="col-12 col-lg-12 text-center">
-        <h2>@Model.Title</h2>
+        <h2 asp-for="@Model.Title"></h2>
     </div>
 }
 
 @{
     var fileName = string.Empty;
 
-    switch (Model.listType)
+    switch (Model.ContentListType.Id)
     {
-        case "{E58A021E-74D2-471E-9EBA-B09616350605}": // White blocks
+        case var id when(id == new Guid("E58A021E-74D2-471E-9EBA-B09616350605")): // White blocks
             fileName = "_clFeatureBlock.cshtml";
             break;
-        case "{92DD658A-76BE-437A-86E4-589DF627FBCC}": // MVP 3 columns
+        case var id when(id == new Guid("92DD658A-76BE-437A-86E4-589DF627FBCC")): // MVP 3 columns
             fileName = "_clMvpYearBlockThreeColumn.cshtml";
             break;
-        case "{F0027CB2-8725-48EA-B85D-EB83E5216E5A}": // MVP 4 columns
+        case var id when(id == new Guid("F0027CB2-8725-48EA-B85D-EB83E5216E5A")): // MVP 4 columns
             fileName = "_clMvpYearBlockFourColumn.cshtml";
             break;
         default:
@@ -27,18 +27,18 @@
     }
 }
 
-@if (Model.limit > 0)
+@*@if (Model.ContentListLimit.Value > 0)
 {
 
-    @foreach (var item in Model.Items.Take(Model.limit))
+    @foreach (var item in Model.Items.Take(Model.ContentListLimit.Value.Value as int))
     {
         @await Html.PartialAsync(fileName, item)
     }
 }
 else
-{
+{*@
     @foreach (var item in Model.Items)
     {
         @await Html.PartialAsync(fileName, item)
     }
-}
+@*}*@
