+++ src/Feature/BasicContent/rendering/Views/Shared/Components/SitecoreComponent/_clFeatureBlock.cshtml
@@ -2,13 +2,12 @@
 
 <div class="col-12 col-lg-6">
     <div class="feature-panel-text with-background no-number">
-        <h2>@Model.ItemTitle</h2>
-        <p>@Model.ItemText</p>
+        <h2 asp-for="@Model.Fields.ItemTitle"></h2>
+        <p asp-for="@Model.Fields.ItemText"></p>
 
-          <!-- UPDATE THIS MODEL! -->
-@*        <a href="@Model.ItemLinkUrl" data-module="cta-button" data-goals="" class="link-arrow" aria-label="Learn more" target="">
-            @Model.ItemLinkDescription
+        <a asp-for="@Model.Fields.ItemLink" data-module="cta-button" data-goals="" class="link-arrow" aria-label="Learn more">
+            @Model.Fields.ItemLink.Value.Text
             <img alt="Learn more" src="images/icon-arrow.png">
         </a>
-*@    </div>
+    </div>
 </div>
