+++ src/Project/MvpSite/items/content/MvpSite/Home.yml
@@ -128,6 +128,12 @@ Languages:
               s:par="ContentListType=%7BE58A021E-74D2-471E-9EBA-B09616350605%7D&amp;ContentListLimit"
               s:ph="/main/column-container-{D2DE8BE0-C911-46EB-9A32-DB8B4863AB2B}-0"
               s:ccb="Clear on publish" />
+            <r
+              uid="{7A931726-B26B-44A0-BCE2-7A1023245601}"
+              s:ccb="Clear on publish"
+              s:ds=""
+              s:par=""
+              s:ph="header-signin-disabled" />
           </d>
         </r>
     - ID: "25bed78c-4957-4165-998a-ca1b52f67497"
@@ -135,7 +141,7 @@ Languages:
       Value: 20200513T045800Z
     - ID: "8cdc337e-a112-42fb-bbb4-4143751e123f"
       Hint: __Revision
-      Value: "b8231b40-cf5c-4665-a767-8d4b3c1dc965"
+      Value: "86729d30-95e6-4c4f-bc4e-37089d948d96"
     - ID: "b1ece12f-b2c9-44e6-92a3-c5dfd6514b59"
       Hint: MenuTitle
       Value: Home
@@ -145,4 +151,4 @@ Languages:
         sitecore\xmc-e2e-admin@sitecore.com
     - ID: "d9cf14b1-fa16-4ba6-9288-e8a174d4d522"
       Hint: __Updated
-      Value: 20220519T021427Z
+      Value: 20220524T002053Z
