+++ src/Project/MvpSite/items/content/MvpSite/Home/Contact.yml
@@ -59,6 +59,12 @@ Languages:
               s:id="{DF26CA0D-A60B-4349-8908-E7A62F7E7272}"
               s:ph="main" />
             <r
+              uid="{BC97B8B5-67B1-4CCA-A391-6B5E7ACCBA32}"
+              s:ccb="Clear on publish"
+              s:ds=""
+              s:par=""
+              s:ph="header-signin-disabled" />
+            <r
               uid="{64CD74AD-6939-48FB-ABB9-33F2403ED031}"
               p:after="*[1=2]" />
           </d>
@@ -68,17 +74,17 @@ Languages:
       Value: 20200604T042308Z
     - ID: "8cdc337e-a112-42fb-bbb4-4143751e123f"
       Hint: __Revision
-      Value: "5ddfbfd0-652c-4b84-9df2-23cb79d12507"
+      Value: "9abae455-f8c0-4c13-9dbd-5cec8f47d78a"
     - ID: "b1ece12f-b2c9-44e6-92a3-c5dfd6514b59"
       Hint: MenuTitle
       Value: Contact
     - ID: "badd9cf9-53e0-4d0c-bcc0-2d784c282f6a"
       Hint: __Updated by
       Value: |
-        sitecore\rob
+        sitecore\xmc-e2e-admin@sitecore.com
     - ID: "d9cf14b1-fa16-4ba6-9288-e8a174d4d522"
       Hint: __Updated
-      Value: 20220406T045741Z
+      Value: 20220524T013917Z
     - ID: "efc0a3c7-6993-47d2-ba85-10b44961916e"
       Hint: IncludeInMenu
       Value: 1
