+++ src/Project/MvpSite/items/content/MvpSite/Shared Content/Content Lists/MVP Categories.yml
@@ -11,13 +11,26 @@ Languages:
     - ID: "25bed78c-4957-4165-998a-ca1b52f67497"
       Hint: __Created
       Value: 20200629T111535Z
+    - ID: "29b266ff-564e-4b11-a10a-70b8d290cd08"
+      Hint: Selected Items
+      Value: |
+        {6770CEE3-D0A6-4F68-94EC-9C8D0B23F8AD}
+        {201315D2-844B-4004-8100-A7244EE24137}
+        {17302488-06E8-4B59-8165-FFECC726C258}
+        {FCD0EB59-694C-46D8-83ED-81C814015EAE}
     - ID: "8cdc337e-a112-42fb-bbb4-4143751e123f"
       Hint: __Revision
-      Value: "d1b76ddb-a2ee-4cf0-b495-265f24cb75e1"
+      Value: "bc7d23f9-bf7a-45c1-99e8-92402a1609b9"
     - ID: "badd9cf9-53e0-4d0c-bcc0-2d784c282f6a"
       Hint: __Updated by
       Value: |
-        sitecore\Admin
+        sitecore\xmc-e2e-admin@sitecore.com
+    - ID: "d40ebb0f-36ff-4035-85f0-412555f4e599"
+      Hint: ContentListType
+      Value: "{E58A021E-74D2-471E-9EBA-B09616350605}"
     - ID: "d9cf14b1-fa16-4ba6-9288-e8a174d4d522"
       Hint: __Updated
-      Value: 20200629T111735Z
+      Value: 20220524T013437Z
+    - ID: "e18a5161-33b5-4652-af07-a19404c2d2ab"
+      Hint: ContentListTitle
+      Value:
