+++ src/Project/MvpSite/items/templates/MvpSite/Homepage/__Standard Values.yml
@@ -18,9 +18,12 @@ SharedFields:
         id="{FE5D7FDF-89C0-4D99-9AA3-B5FBD009C9F3}"
         l="{1AE49D12-8DA1-4FBD-92FF-887B01C02BB5}">
         <r
+          ds=""
           id="{00147A0F-476B-4C99-8A10-E6B4FCFCECE0}"
-          ph="header-signin"
-          uid="{9DD9C0B4-97B5-4F10-B6E8-C30CC5EE32A5}" />
+          par=""
+          ph="header-signin-disabled"
+          uid="{9DD9C0B4-97B5-4F10-B6E8-C30CC5EE32A5}"
+          ccb="Clear on publish" />
       </d>
     </r>
 Languages:
@@ -41,11 +44,11 @@ Languages:
         sitecore\Admin
     - ID: "8cdc337e-a112-42fb-bbb4-4143751e123f"
       Hint: __Revision
-      Value: "317422ab-aa5a-405a-94a2-a3dd9d668eb3"
+      Value: "8c19dc75-f7cd-44ea-b514-fe5ce7a0ad32"
     - ID: "badd9cf9-53e0-4d0c-bcc0-2d784c282f6a"
       Hint: __Updated by
       Value: |
-        sitecore\Admin
+        sitecore\xmc-e2e-admin@sitecore.com
     - ID: "d9cf14b1-fa16-4ba6-9288-e8a174d4d522"
       Hint: __Updated
-      Value: 20220303T112904Z
+      Value: 20220524T013824Z
